+++ lib/http/mod.rs
@@ -6,4 +6,5 @@ pub mod header;
 pub mod method;
 pub mod multipart;
 pub mod query;
+pub mod server_sent_events;
 pub mod status;